+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 158:
 * Tidy up end_of_stream javadoc
 * Tidy up websocket docs
 * Examples set reuse_address(true)
+* Advanced servers support clean shutdown via SIGINT or SIGTERM
 
 --------------------------------------------------------------------------------